Align input options for plotting empty reads
Input options for showing/including empty reads in tabulateEpibed and
plotEpiread now have similar naming scheme so that the TRUE/FALSE inputs
match between the two when running in NOMe mode. This isn't a very
explanative message, so see the note/comment that was added in
plotEpiread with this commit for why this was done.